+++ src/TableParser/HTML/HTMLStringTableParser.php
@@ -14,7 +14,45 @@ use LibXMLError;
 /**
  * Parses an HTML string for a <table> element into a TableNode. The table must have a single row
  * in the <thead> and a <tbody> element (though this does not have to contain any rows). You can
- * skip additional rows in the table by marking up like <tr data-behat-table="ignore">
+ * skip additional rows in the table by marking up like <tr data-behat-table="ignore">.
+ *
+ * If you need to prefix cell content with some string for behat (for example to differentiate
+ * two columns that would be visually different to a human user) add a data-behat-table-prefix
+ * attribute.
+ *
+ *   <table>
+ *     <thead>
+ *       <tr data-behat-table="ignore">
+ *         <th></th>
+ *         <th colspan=2>Timings</th>
+ *         <th></th>
+ *       </tr>
+ *       <tr>
+ *         <th>Category</th>
+ *         <th data-behat-table-prefix="Timings -">Before</th>
+ *         <th data-behat-table-prefix="Timings -">After</th>
+ *         <th>Count</th>
+ *       </tr>
+ *     </thead>
+ *     <tbody>
+ *       <tr>
+ *         <th>Thing</th>
+ *         <td>12.3</td>
+ *         <td>14.2</td>
+ *         <td>12</td>
+ *       </tr>
+ *       <tr>
+ *         <th colspan="3">Total</th>
+ *         <td>12</td>
+ *       </tr>
+ *     </tbody>
+ *   </table>
+ *
+ * becomes:
+ *
+ *   | Category | Timings - Before | Timings - After | Count |
+ *   | Thing    | 12.3             | 14.2            | 12    |
+ *   | Total    | ...              | ...             | 12    |
  *
  * @package Ingenerator\BehatTableAssert\TableParser\HTML
  */
@@ -144,22 +182,27 @@ class HTMLStringTableParser
                 continue;
             }
 
-            $rows[] = $this->findChildTextValues($row);
+            $rows[] = $this->findCellTextValues($row);
         }
         return $rows;
     }
 
     /**
-     * @param \SimpleXMLElement $element
+     * @param \SimpleXMLElement $table_row
      *
      * @return string[]
      */
-    protected function findChildTextValues(\SimpleXMLElement $element)
+    protected function findCellTextValues(\SimpleXMLElement $table_row)
     {
         $row = [];
-        foreach ($element->children() as $child) {
+        foreach ($table_row->children() as $child) {
             /** @var \SimpleXMLElement $child */
-            $row[] = trim(preg_replace('/\s+/', ' ', dom_import_simplexml($child)->textContent));
+            $text = trim(preg_replace('/\s+/', ' ', dom_import_simplexml($child)->textContent));
+
+            if ($prefix = (string) $child['data-behat-table-prefix']) {
+                $text = $prefix.' '.$text;
+            }
+            $row[] = $text;
 
             $colspan = (int) $child['colspan'];
             for ($i = 1; $i < $colspan; $i++) {