+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 * [FEATURE] Parse multiple colspan columns in HTML tables as `...`
 * [FEATURE] Support skipping over presentational rows in HTML tables with data attribute
+* [FEATURE] Optionally prefix HTML table cell values with data attribute
 
 ## 1.0.0 / 2015-10-09